CODE ERRORS SUBMISSION ASSIGNMENT
fixed the following errors:

The Light-blue background color is not displaying correctly.

The Heading is not displaying properly.

The Title element is not functioning correctly.

The Paragraph, "I am a bold student of The Tech Academy" needs to be bolded.

The URL to The Tech Academy's website is not functioning correctly.

The Tech Academy Logo is out of proportion. It should be resized with a width of 600px and a height of 300px.

The Tech Academy Logo needs to display a red, dashed border, two pixels in thickness.

The Heading 3 element, "Have you seen The Tech Academy's Student Orientation video?" is not being centered correctly.

The embedded Youtube video isn't displaying.